requirements.txt
streamlit
PyPDF2
groq
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
faiss-cpu
pandas
python-dotenv
